requirements.txt
fastapi[all]
huggingface-hub
idna
imageio
intel-openmp
Jinja2
joblib
MarkupSafe
matplotlib
matplotlib-inline
mkl
mpmath
munch
nest-asyncio
networkx
numpy
opencv-python-headless
packaging
pandas
parso
pillow
platformdirs
pretrainedmodels
prompt-toolkit
psutil
pure-eval
pydantic
pydantic_core
Pygments
pyparsing
python-dateutil
pytz
pywin32
PyYAML
pyzmq
regex
requests
safetensors
scikit-image
scikit-learn
scipy
segmentation-models-pytorch
transformers
